+++ Sources/Tests/StoryboardMappingTestCase.swift
@@ -8,7 +8,6 @@
 
 import UIKit
 import XCTest
-import DTModelStorage
 @testable import DTTableViewManager
 
 class StoryboardMappingTestCase: XCTestCase {